refactor: external command
